backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
textstat==0.7.3
PyPDF2==3.0.1
